Kapitel18/mcp_py/pyproject.toml
[project]
name = "bdb-ki-mcp"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.11.3",
    "jinja2>=3.1.6",
    "langchain>=0.3.27",
    "langchain-mcp>=0.2.1",
    "langchain-ollama>=0.3.7",
    "langchain-openai>=0.3.31",
    "langgraph>=0.6.6",
    "pandas>=2.3.2",
]
